refactor(model): Move inverse profit and payout calculation
They belong in the `model::payout_curve::inverse` module, for they are
specific to that payout curve.
