+++ daemon/src/projection.rs
@@ -14,8 +14,6 @@ use futures::StreamExt;
 use itertools::Itertools;
 use maia_core::TransactionExt;
 use model::calculate_long_liquidation_price;
-use model::calculate_profit;
-use model::calculate_profit_at_price;
 use model::calculate_short_liquidation_price;
 use model::long_and_short_leverage;
 use model::market_closing_price;
@@ -584,7 +582,7 @@ impl Cfd {
 
         // If we have a dedicated closing price, use that one.
         if let Some(payout) = self.aggregated.clone().payout(self.role) {
-            let (profit_btc, profit_percent) = calculate_profit(payout, self.margin);
+            let (profit_btc, profit_percent) = inverse::calculate_profit(payout, self.margin);
 
             return Self {
                 payout: Some(payout),
@@ -632,7 +630,7 @@ impl Cfd {
         let (long_leverage, short_leverage) =
             long_and_short_leverage(self.leverage_taker, self.role, self.position);
 
-        let (profit_btc, profit_percent, payout) = match calculate_profit_at_price(
+        let (profit_btc, profit_percent, payout) = match inverse::calculate_profit_at_price(
             self.initial_price,
             closing_price,
             self.quantity,
@@ -902,7 +900,7 @@ impl sqlite_db::ClosedCfdAggregate for Cfd {
             (CfdDetails { tx_url_list }, price, payout, state)
         };
 
-        let (profit_btc, profit_percent) = calculate_profit(payout.inner(), margin);
+        let (profit_btc, profit_percent) = inverse::calculate_profit(payout.inner(), margin);
 
         // there are no events to apply at this stage for closed CFDs,
         // which is why this field is mostly ignored
